src/third_party/blink/web_tests/http/tests/loading/dont-preload-non-img-srcset-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didReceiveTitle: 
frame "<!--framePath //<!--frame0-->-->" - DidStartNavigation
frame "<!--framePath //<!--frame0-->-->" - ReadyToCommitNavigation
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didReceiveTitle: 
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
main frame - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didReceiveTitle: 
frame "<!--framePath //<!--frame0-->-->" - DidStartNavigation
frame "<!--framePath //<!--frame0-->-->" - ReadyToCommitNavigation
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didReceiveTitle: 
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
main frame - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded('resources/base-image1.png'); is true
PASS internals.isPreloaded('resources/empty.html'); is false
PASS internals.isPreloaded('resources/empty.js'); is false
PASS successfullyParsed is true

TEST COMPLETE

This test passes if the image is preloaded, but none of the non-img elements with a srcset attribute is.
 
